requirements.txt
django-tinymce4-lite==1.7.5
django-phone-field==1.8.0
django-reset-migrations==0.3.1
Pillow==6.1.0